Removing an inappropriate try...catch.
The library was swallowing AccessViolationExceptions and logging them to
stderr (a remnant from before it was a library, I hope). Clearly this is
inappropriate behavior for a class library to be doing, especially one
that originates in Windows, where most users of the library (if it even
works) will probably want to write graphical or Web applications (which
lack a stderr). I digress... Begone, try...catch!